# Improve user /lists page performance by remove top subjects