This is making a lot of expensive requests to solr, and providing very minimal value.

<!-- What issue does this PR close? -->
Closes #3896

<!-- What does this PR achieve? [feature|hotfix|fix|refactor] -->

### Technical
<!-- What should be noted about the implementation? -->

### Testing
- [x] http://staging.openlibrary.org/people/Hermgirl/lists 11s → 2s
- [x] http://staging.openlibrary.org/people/seabelis/lists 45s → 6s
 
### Screenshot
| Before | After | 
| --- | --- |
| ![image](https://user-images.githubusercontent.com/6251786/110182375-a33da200-7dda-11eb-95ac-803fc90657cc.png)  | ![image](https://user-images.githubusercontent.com/6251786/110182315-8c974b00-7dda-11eb-937a-96d9ab3184f1.png) |


### Stakeholders
@mekarpeles  @seabelis