@mekarpeles 2021-03-10
Please do courtesy tag on PRs impacting product so I have a chance to :eyes:, these changes seem fine given the performance characteristics